feat: Reaction型を新しく #10
