+++ mipac/models/__init__.py
@@ -2,8 +2,9 @@ from .chat import Chat
 from .drive import File, Folder, Properties
 from .emoji import Emoji
 from .instance import Instance, InstanceMeta
-from .note import Follow, Header, Note, NoteReaction, Poll, Reaction, Renote
+from .note import Follow, Header, Note, NoteReaction, Poll, Renote
 from .user import Followee, FollowRequest, User
+from .notification import Reaction
 
 __all__ = (
     'Chat',
